conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py36_0
astor=0.7.1=py36_0
backcall=0.1.0=py36_0
blas=1.0=mkl
bzip2=1.0.6=h14c3975_5
c-ares=1.15.0=h7b6447c_1
ca-certificates=2019.1.23=0
cairo=1.14.12=h8948797_3
certifi=2019.3.9=py36_0
cudatoolkit=10.0.130=0
cudnn=7.3.1=cuda10.0_0
cupti=10.0.130=0
cycler=0.10.0=py36_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py36_1
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
gast=0.2.2=py36_0
glib=2.56.2=hd408876_0
graphite2=1.3.13=h23475e2_0
grpcio=1.16.1=py36hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py36h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
imageio=2.5.0=py36_0
intel-openmp=2019.3=199
ipython=7.5.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
jasper=2.0.14=h07fcdf6_1
jedi=0.13.3=py36_0
jpeg=9b=h024ee3a_2
keras=2.2.4=0
keras-applications=1.0.7=py_0
keras-base=2.2.4=py36_0
keras-preprocessing=1.0.9=py_0
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.7.1=hd408876_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
markdown=3.1=py36_0
matplotlib=3.0.3=py36h5429711_0
mkl=2019.3=199
mkl_fft=1.0.12=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
mock=2.0.0=py36_0
mpi4py=2.0.0=py36_2
mpich2=1.4.1p1=0
ncurses=6.1=he6710b0_1
numpy=1.16.3=py36h7e9f1db_0
numpy-base=1.16.3=py36hde5b4d6_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1b=h7b6447c_1
pandas=0.24.2=py36he6710b0_0
parso=0.4.0=py_0
pbr=5.1.3=py_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=py36_0
pillow=6.0.0=py36h34e0f95_0
pip=19.0.3=py36_0
pixman=0.38.0=h7b6447c_0
prompt_toolkit=2.0.9=py36_0
protobuf=3.7.1=py36he6710b0_0
psutil=5.6.2=py36h7b6447c_0
ptyprocess=0.6.0=py36_0
py-opencv=3.4.2=py36hb342d67_1
pygments=2.4.0=py_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py36_0
pytz=2019.1=py_0
pyyaml=5.1=py36h7b6447c_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
scipy=1.2.1=py36h7c811a0_0
setuptools=40.8.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.27.2=h7b6447c_0
tensorboard=1.13.1=py36hf484d3e_0
tensorflow=1.13.1=gpu_py36h3991807_0
tensorflow-base=1.13.1=gpu_py36h8d69cac_0
tensorflow-estimator=1.13.0=py_0
tensorflow-gpu=1.13.1=h0d30ee6_0
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
tornado=6.0.2=py36h7b6447c_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36_0
werkzeug=0.15.2=py_0
wheel=0.33.1=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
